+++ irc/server.go
@@ -1677,3 +1677,38 @@ func whowasHandler(server *Server, client *Client, msg ircmsg.IrcMessage) bool {
 	}
 	return false
 }
+
+// LUSERS [ <mask> [ <target> ] ]
+func lusersHandler(server *Server, client *Client, msg ircmsg.IrcMessage) bool {
+	//TODO(vegax87) Fix Unknown connections and additional parameters
+
+	var totalcount int
+	var invisiblecount int
+	var opercount int
+	var chancount int
+
+	server.clients.ByNickMutex.RLock()
+	defer server.clients.ByNickMutex.RUnlock()
+
+	for _, onlineusers := range server.clients.ByNick {
+		totalcount += 1
+		if onlineusers.flags[Invisible] {
+			invisiblecount += 1
+		}
+		if onlineusers.flags[Operator] {
+			opercount += 1
+		}
+	}
+
+	for chans := range server.channels {
+		//Little hack just to avoid "variable declared but not used" error
+		_ = chans
+		chancount += 1
+	}
+	client.Send(nil, server.name, RPL_LUSERCLIENT, client.nick, fmt.Sprintf("There are %d users and %d invisible on %d server(s)", totalcount, invisiblecount, 1))
+	client.Send(nil, server.name, RPL_LUSEROP, client.nick, fmt.Sprintf("%d operators online", opercount))
+	client.Send(nil, server.name, RPL_LUSERUNKNOWN, client.nick, "X unknown connection(s)")
+	client.Send(nil, server.name, RPL_LUSERCHANNELS, client.nick, fmt.Sprintf("%d channels formed", chancount))
+	client.Send(nil, server.name, RPL_LUSERME, client.nick, fmt.Sprintf("I have %d clients and %d servers", totalcount, 1))
+	return false
+}