+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New release of Oragono!
 * Added ARM build (for Raspberry PIs and similar).
 * Added automated connection throttling! See the new `connection-throttling` section in the config.
 * Added `KLINE` and `UNKLINE` commands. Complementing `DLINE`'s per-IP and per-network bans, this lets you ban masks from the server.
+* Added LUSERS command. It works for a single server for now (by @vegax87)
 
 ### Changed
 * Changed casemapping from "rfc7700" to "rfc7613", to match new draft spec.